pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "multicloud-storage"
description = ""
version = "0.7.0"
readme = "README.md"
authors = ["Adam Williams <adam.williams@onxmaps.com>"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.7"
google-cloud-storage = "^1.41.0"
minio = "^7.1.0"
python-dotenv = "^0.18.0"
typing-extensions = "3.10.0.0"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 10
experimental-string-processing = true
